Fix translation strings, move PauseTime class to IgnoredUpdatesDatabase
